requirements.txt
Babel
eventlet
iso8601
oslo.config
pbr
pygerrit
python-jenkins
six
stevedore
PyYAML